feat : Engine modularization
